# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.5.10",
    "title": "META API",
    "description": "API文件"
  },
  "host": "six-northnode.onrender.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "jwt token"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/screens/moviePlayDate": {
      "post": {
        "tags": [
          "Screen"
        ],
        "summary": "所選擇電影可上映的日期",
        "description": "傳入screenId，來得到該電影的可訂票日期",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "screenId": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "回傳範例資料",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "自行填寫"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date": {
                        "type": "string",
                        "example": "5/20"
                      },
                      "screenId": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/screens/moviePlayTime": {
      "post": {
        "tags": [
          "Screen"
        ],
        "summary": "所選擇電影可上映的日期",
        "description": "傳入screenId，來得到該電影的可訂票日期",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "screenId": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "回傳範例資料",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "自行填寫"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date": {
                        "type": "string",
                        "example": "5/20"
                      },
                      "screenId": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/screens/moviePlaySeats": {
      "post": {
        "tags": [
          "Screen"
        ],
        "summary": "所選擇電影可上映的日期",
        "description": "傳入screenId，來得到該電影的可訂票日期",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "screenId": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "回傳範例資料",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "自行填寫"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date": {
                        "type": "string",
                        "example": "5/20"
                      },
                      "screenId": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/screens/screenAdd": {
      "post": {
        "tags": [
          "Screen"
        ],
        "summary": "所選擇電影可上映的日期",
        "description": "傳入screenId，來得到該電影的可訂票日期",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "screenId": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "回傳範例資料",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "自行填寫"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date": {
                        "type": "string",
                        "example": "5/20"
                      },
                      "screenId": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
...
}